+++ CHANGELOG.md
@@ -4,7 +4,11 @@ Date format: (YYYY-MM-DD)
 ## Next release
 ### Supported MC versions: xxx
 
-## v2.9.0 Release (TBA)
+## v2.9.1 Release (TBA)
+### Supported MC versions: 1.18.1
+
+
+## v2.9.0 Release (2022-01-04)
 ### Supported MC versions: 1.18.1
 * Bumped the Bukkit dependency to Spigot 1.18.1. This version is not compatible with older Bukkit versions.
 * Bumped the ProtocolLib dependency to version 4.8.0.